Add a massive number of things.
This commit is unconscientiously large. It adds a new MainActivity screen, which is an options layer between  the rest of the app. It adds the settings menu and the rest of the settings features. It adds the Google Maps API and a map on the navigate screen, with markers to show the route. It also fixes and tweaks a bunch of bugs, and changes the layout and layout colors.